backend/requirements.txt
# FastAPI and web server
fastapi>=0.104.1
setuptools>=68.0.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Vector Store and Embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2

# HTTP Requests
requests>=2.31.0
aiohttp>=3.9.1

# Data Processing
beautifulsoup4>=4.12.2
pypdf>=3.17.1

# Image Processing & OCR
pytesseract>=0.3.10
python-multipart>=0.0.6
Pillow>=10.0.0

# Language Detection
langdetect>=1.0.9

# Additional
numpy>=1.26.0
groq>=0.4.2
